fix: use getInitialProps to load opengraph data
This allows us to completely prevent making any calls to the server
on client-side navigation, which would do nothing anyway, since
we already have the data in React Query's cache.

Using getInitialProps we can return any public watchlist info to
the client on first render. A downside to this is that we have
to check for error codes on the client now